feat: add way to deserialize cluster definition
Each cluster contains a "cluster definition" file. This special file is
housed at the top-level of the cluster. For now I got the
deserialization logic to work, but I still need to implement the logic
that extracts the file from the top-level and parses it in one shot when
creating a new cluster.

Signed-off-by: Jason Pena <jasonpena@awkless.com>