feat(main): add `init` command
The new `init` command has been added to the oxidot's CLI. This command
allows the user to initialize a new cluster within the cluster store. It
automatically stages and commits a usage cluster definition for the user
to later utilize for future command implementations.

For now, the user can set the description, url, and work tree alias for
the initial cluster definition that the `init` command creates for them.

Signed-off-by: Jason Pena <jasonpena@awkless.com>