fix: avoid live-config limitation in `Cluster::try_new_open`
Okay, just found out that libgit2 cannot allow for a mutable config
handle when a key does not yet exist. Thus, I have to use
`git2::Config::get_entry` as a workaround to actually obtain the
contents of a configuration value.

Apparently, when a key is missing, libgit2 tries to resolve it per
configuration level dynamically, which needs each config to be
immutable. However, `git2::Config::get_str` allows for a mutable string,
which conflicts with the key config resolution that libgit2 uses[1].
Thus, `git2::Config::get_str` can only be used on a snapshot of the
repository configuration. Very weird quirk, but at least the code now
works.

[1]: https://docs.rs/git2/latest/git2/struct.Config.html#method.get_str

Signed-off-by: Jason Pena <jasonpena@awkless.com>