feat: add `Store::resolve_dependencies`
Now the `Store` API can resolve the dependencies of a cluster entry.
Just using an iterative variant of DFS to traverse the dependency tree.
Once a dependency is found to not yet exist, but was declared, then
`resolve_dependencies` will just clone it based on the `url` field given
(if valid).

Signed-off-by: Jason Pena <jasonpena@awkless.com>