fix: ensure that sparse checkout file is created
Simple fix that ensures that when a new `SparseCheckout` object is made,
that the sparse checkout file is created for a given cluster. I also
made errors more descriptive for `SparseCheckout`.

I also added `Store::get` to easily get a cluster from the cluster
store. Probably should have added this as a separate commit, but
oh well.

Signed-off-by: Jason Pena <jasonpena@awkless.com>