fix: sparsity rules not being on their own line
Original code in the `SparseCheckout` API would shove all existing rules
onto the same line. This is not the correct behavior. Each rule being
written to the sparse checkout file must be on its own line. Thus, the
changes seen in this code fixes this issue. I also simplified the read
and write logic for file handling making it more readable.

Signed-off-by: Jason Pena <jasonpena@awkless.com>