chore: add `ignore` crate
I need some way to parse and match gitignore style rules. The idea is to
use this crate to figure out if a given cluster is deployed by checking
if the paths match a given sparsity rule. This crate will allow me to do
that by giving me access to the gitignore syntax, because sparse
checkout use the same syntax but the logic is inverted!

Signed-off-by: Jason Pena <jasonpena@awkless.com>