feat: add `Store::list_fully`
Fully list all details of each cluster in the cluster store. Shows if
clusters are deployed, the name of each cluster, and what their
work tree aliases are.

I also fixed the logic in `Cluster::is_deployed`. Originally, it only
checked if paths existed, but did not check if the paths actually match
the current sparsity rules. So if two clusters had a file with the same
name, and only one of them was deployed, then the `Cluster::is_deployed`
method would consider both to be deployed.

I fixed the `Cluster::is_deployed` logic by using the `ignore` crate.
Sparse checkout uses the same syntax as gitignore syntax. However, the
semantics are reversed! Gitignore syntax excludes tracked files that
match the rules, but sparse checkout syntax includes them. So in order
for me to check if a given path matches a sparsity rule using the
`ignore` crate API, then I needed to negate/invert the sparsity rules
_before_ passing them to the `ignore` crate. This approach seems to
work, granted it is a little convoluted.

I also made sure to make `Cluster::is_deployed` utilize a first
occurrence search approach. In other words, once the method finds a
matching and existing path, it immediately returns true. Otherwise,
it will continue iterating through each path with each rule. The
worst case time complexity becomes O(MxN) where M is the number of
paths, and N is the number of rules to match. Not great, but works.

Signed-off-by: Jason Pena <jasonpena@awkless.com>