fix: keep sparse rules and index in sync after interactive git call
The git shortcut command would always fail when trying to add new files
into a cluster, i.e., `oxidot <cluster_name> add [<files>]...`. This is
because any new files are not apart of the sparse checkout configuration
file definition. The only way to fix this is to append the `--sparse`
flag for git commands to modify the index.

Thus, the fix is to add that `--sparse` flag when needed, and update
the sparse checkout configuration file to sync sparse checkout with the
newly updated index. Now the git shortcut command works the way we
want it to!

Signed-off-by: Jason Pena <jasonpena@awkless.com>