fix: handle directory patterns for sparsity rules
The original logic I used to handle sparsity patterns meant to match
directories was horribly broken. The inversion logic for the `ignore`
crate API was not correct, leading to additional rules being written
into the sparsity configuration file.

The new logic detects negation syntax and properly inverts it, with
directory syntax finally being supported properly. I also updated the
logic for syncing the sparse checkout rules with the index to only
write new rules if and only if the index itself actually changes.

Signed-off-by: Jason Pena <jasonpena@awkless.com>