refactor!: add `oxidot::cluster::domain::Cluster`
This is the beginning of along process of breaking up the infamous `Custer` god
class that was used to implement the command set for the project in the
early stages of development.

My current idea is to use the new APIs provided `oxidot::cluster::sparse::*`,
and the configuration layout `oxidot::config::ClusterDefinition` to form
the basis of the new `Cluster` type being developed. I will eventually
define a separate type to handle the construction of a `Cluster` in
order to more properly wrap the libgit2 API away from the `Cluster` API
itself.

Deployment logic will also be defined separately. I have yet to refactor
deployment logic, so it has been left out for now.

Signed-off-by: Jason Pena <jasonpena@awkless.com>