feat!: specify branch settings through new `ClusterRemote` setting
I want to introduce the ability for oxidot to switch to a different
branch when cloning a dependency. Thus, I changed the `url` setting to
use the new `remote` setting where the user can specify the URL and
branch that they want a dependency to use.

Now the `Store::clone_cluster` can clone with the target branch of
dependencies. The CLI now offers the ability to initialize a new cluster
with a default target branch besides the default one.

Breaking-changes: URL settings must be specified through `ClusterRemote`.
Signed-off-by: Jason Pena <jasonpena@awkless.com>